Command/flask-enrollment-backend/requirements.txt
###############
# Common:
###############
joblib==1.0.1
scikit-learn==0.24.2
fpdf==1.7.2
PyPDF2==1.26.0

###############
# SpeakerID:
###############
## vosk and scikit learn stuff
sklearn-porter @ git+https://github.com/null0nil/sklearn-porter@746a8e7985791e3713c8037a2ce99cd248a15164
numpy==1.19.5
pandas==1.2.4
PyAudio==0.2.11
pandarallel==1.5.2
vosk==0.3.17

###############
# FaceID:
###############
tensorflow==2.6.0
opencv-python==4.5.2.52
opencv-contrib-python==4.5.2.52
imutils==0.5.4
Pillow==8.2.0
Keras==2.6.0
mtcnn==0.1.0
matplotlib==3.4.1
scikit-image==0.18.1
scanf==1.5.2
libsvm-official==3.25.0

###############
# MISC:
###############
# We need to use exclude_pattern to prevent watchdog from being too aggressive
Werkzeug==2.0.1
flask-swagger-ui==3.36.0
flask-praetorian==1.0.0
boto==2.49.0
